Adjusted test scene
Made the font for the dialogue box the Hendrix font with Hendrix colors, Made a separate text box for the name of the character to display and moved these to the bottom of the canvas. I also added the 8 pixel version of dwr as the background to see how it looks.