Created Resume Script/UI object and altered both GameManager and Player
Created the barebones of a Resume Script that for now only creates a reference to a compentency.
With that in mind, I altered Gamemanager to more efficiently use the player object, and in theory should keep the object as a constant data source for competency. I also made some changes to type text to start writing in logic for the text to print in full while typing and then move on to the next dialog on another click.